- Se agrego la configuracion de swagger UI - Se agrego las anotaciones de swagger en los controladores - Se agrego las anotaciones de swagger y de lombok en las clases entity - Se cambio el id de Product por un Long - Se agregaron la documentacion de api-doc
