+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
     command: --default-authentication-plugin=mysql_native_password
     restart: always
     ports:
-      - 3307:3306
+      - 3308:3306
     environment:
       MYSQL_ROOT_PASSWORD: root
       MYSQL_DATABASE: pedidos
@@ -19,7 +19,7 @@ services:
       dockerfile: Dockerfile
     container_name: pedidos-app
     ports:
-      - 8080:8080
+      - 8081:8080
     depends_on:
       - pedidos-mysql
     links:
