+++ src/main/java/com/ma/pedido/config/SwaggerConfig.java
@@ -1,0 +1,114 @@
+package com.ma.pedido.config;
+
+import com.ma.pedido.controllers.PizzeriaController;
+import com.ma.pedido.controllers.ProductController;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurationSupport;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.builders.ResponseMessageBuilder;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
+import springfox.documentation.service.ResponseMessage;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger.web.*;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Configuration
+@EnableSwagger2
+@ComponentScan(basePackageClasses = {
+        PizzeriaController.class,
+        ProductController.class
+})
+public class SwaggerConfig extends WebMvcConfigurationSupport {
+    /** Para configurar Swagger en una aplicacion se debe:
+     * - Agregar la dependencia de SpringFox Swagger2
+     * - Usar la anotacion @EnableSwagger2 en la clase Main
+     * Con esta configuracion ya se puede ver la informacion de las API en formato JSON pegandole al endpoint
+     * http://localhost:8080/v2/api-docs
+     */
+
+    /**
+     * Para configurar la UI de Swagger
+     * hay que descargar la dependencia de SpringFox-UI (tiene que tener la misma version que SpringFox)
+     * Para customizer mas la UI se puede usar una clase config de swagger y anotaciones en la clases Controllers y Entitys
+     * http://localhost:8080/swagger-ui.html
+     */
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.any())
+                .build()
+                .globalResponseMessage(RequestMethod.GET, commonMessaje());
+    }
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("swagger-ui.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
+
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+    }
+
+    private List<ResponseMessage> commonMessaje() {
+        List<ResponseMessage> responseMessages = new ArrayList<>();
+        responseMessages.add(new ResponseMessageBuilder().code(HttpStatus.BAD_REQUEST.value())
+                .message("").build());
+        return responseMessages;
+    }
+
+    private ApiInfo getApiInfo() {
+
+        return new ApiInfoBuilder()
+                .title("Swagger API Doc")
+                .description("More description about the API")
+                .version("1.0.0")
+                .build();
+    }
+
+    private ApiInfo metaData() {
+        return new ApiInfoBuilder()
+                .title("Spring Boot REST API delivery application.")
+                .description("\"Spring Boot REST API for Online delivery app\"")
+                .version("1.0.0")
+                .license("Apache License Version 2.0")
+                .licenseUrl("https://www.apache.org/licenses/LICENSE-2.0\"")
+                .contact(new Contact("Ignacio Luna", "http://springframework/", "ignacioaniball@gmail.com"))
+                .build();
+    }
+
+    @Bean
+    UiConfiguration uiConfig() {
+        return UiConfigurationBuilder.builder()
+                .deepLinking(true)
+                .displayOperationId(false)
+                .defaultModelsExpandDepth(1)
+                .defaultModelExpandDepth(1)
+                .defaultModelRendering(ModelRendering.EXAMPLE)
+                .displayRequestDuration(false)
+                .docExpansion(DocExpansion.NONE)
+                .filter(false)
+                .maxDisplayedTags(null)
+                .operationsSorter(OperationsSorter.ALPHA)
+                .showExtensions(false)
+                .tagsSorter(TagsSorter.ALPHA)
+                .supportedSubmitMethods(UiConfiguration.Constants.DEFAULT_SUBMIT_METHODS)
+                .validatorUrl(null)
+                .build();
+    }
+}
+
