--- src/main/java/com/ma/pedido/model/dao/OrderDetailDao.java
@@ -1,9 +1,0 @@
-package com.ma.pedido.model.dao;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.ma.pedido.model.entity.OrderDetail;
-
-public interface OrderDetailDao extends CrudRepository<OrderDetail, Long>{
-
-}
