--- src/main/java/com/ma/pedido/model/dto/ProductDto.java
@@ -1,50 +1,0 @@
-package com.ma.pedido.model.dto;
-
-public class ProductDto {
-
-    private String id;
-    private String nombre;
-    private String descripcionCorta;
-    private String descripcionLarga;
-    private Double precioUnitario;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getDescripcionCorta() {
-        return descripcionCorta;
-    }
-
-    public void setDescripcionCorta(String descripcionCorta) {
-        this.descripcionCorta = descripcionCorta;
-    }
-
-    public String getDescripcionLarga() {
-        return descripcionLarga;
-    }
-
-    public void setDescripcionLarga(String descripcionLarga) {
-        this.descripcionLarga = descripcionLarga;
-    }
-
-    public Double getPrecioUnitario() {
-        return precioUnitario;
-    }
-
-    public void setPrecioUnitario(Double precioUnitario) {
-        this.precioUnitario = precioUnitario;
-    }
-}
