--- src/main/java/com/ma/pedido/model/entity/OrderDetail.java
@@ -1,71 +1,0 @@
-package com.ma.pedido.model.entity;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotEmpty;
-import java.io.Serializable;
-
-@Entity
-@Table(name = "pedido_detalle")
-public class OrderDetail implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "pedido_detalle_id")
-    private Long pedidoDetalleId;
-    @ManyToOne(cascade = CascadeType.ALL, targetEntity = Order.class)
-    @JoinColumn(name = "pedido_cabecera_id")
-    private Order order;
-    @OneToOne(targetEntity = Product.class, fetch = FetchType.LAZY)
-    @JoinColumn(name = "producto_id")
-    private Product producto;
-    @NotEmpty
-    private String productId;
-    @NotEmpty
-    @Column(name = "cantidad")
-    private Integer cantidad;
-    @NotEmpty
-    @Column(name = "importe")
-    private Double importe;
-
-    public Order getOrder() {
-        return order;
-    }
-
-    public void setOrder(Order order) {
-        this.order = order;
-    }
-
-    public Product getProducto() {
-        return producto;
-    }
-
-    public void setProducto(Product producto) {
-        this.producto = producto;
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public void setProductId(String productId) {
-        this.productId = productId;
-    }
-
-    public Integer getCantidad() {
-        return cantidad;
-    }
-
-    public void setCantidad(Integer cantidad) {
-        this.cantidad = cantidad;
-    }
-
-    public Double getImporte() {
-        return importe;
-    }
-
-    public void setImporte(Double importe) {
-        this.importe = importe;
-    }
-}
