--- src/main/java/com/ma/pedido/model/request/OrderDetailRequest.java
@@ -1,34 +1,0 @@
-package com.ma.pedido.model.request;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "producto",
-        "cantidad"
-})
-public class OrderDetailRequest {
-
-    @JsonProperty("producto")
-    private String producto;
-    @JsonProperty("cantidad")
-    private Integer cantidad;
-
-    public String getProducto() {
-        return producto;
-    }
-
-    public void setProducto(String producto) {
-        this.producto = producto;
-    }
-
-    public Integer getCantidad() {
-        return cantidad;
-    }
-
-    public void setCantidad(Integer cantidad) {
-        this.cantidad = cantidad;
-    }
-}
