+++ src/main/java/com/ma/pedido/models/dao/OrderDetailDao.java
@@ -1,0 +1,9 @@
+package com.ma.pedido.models.dao;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.ma.pedido.models.entity.OrderDetail;
+
+public interface OrderDetailDao extends CrudRepository<OrderDetail, Long>{
+
+}
