--- src/main/java/com/ma/pedido/service/jpa/OrderDetailService.java
@@ -1,9 +1,0 @@
-package com.ma.pedido.service.jpa;
-
-import com.ma.pedido.model.entity.OrderDetail;
-
-public interface OrderDetailService {
-
-    public void save(OrderDetail orderDetail);
-
-}
