+++ src/main/java/com/ma/pedido/services/PizzeriaService.java
@@ -1,0 +1,16 @@
+package com.ma.pedido.services;
+
+import com.ma.pedido.models.request.OrderRequest;
+import com.ma.pedido.models.response.OrderResponse;
+
+import java.util.List;
+
+public interface PizzeriaService {
+
+    public OrderResponse createOrder(OrderRequest pedido);
+
+    public List getOrderList(String fecha);
+
+    public List getActiveOrderList(String fecha);
+
+}
