+++ src/main/java/com/ma/pedido/services/jpa/OrderDetailService.java
@@ -1,0 +1,9 @@
+package com.ma.pedido.services.jpa;
+
+import com.ma.pedido.models.entity.OrderDetail;
+
+public interface OrderDetailService {
+
+    public void save(OrderDetail orderDetail);
+
+}
