+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 #Database
 # suppress inspection "SpellCheckingInspection"
-spring.datasource.url=jdbc:mysql://localhost:3307/pedidos?user=root&password=root
+spring.datasource.url=jdbc:mysql://localhost:3308/pedidos?user=root&password=root
 spring.datasourcer.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
@@ -8,9 +8,16 @@ spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.dll-auto=create-drop
 
+server.port=8080
+
 #Logging
 logging.level.org.hibernate.SQL=debug
 logging.level.org.springframework.web=DEBUG
 
+
+
 order.discount=30
 order.amountProduct=3
+
+logging.level.org.springframework.web.filter.CommonsRequestLoggingFilter=DEBUG
+
