--- src/main/resources/data.sql
@@ -1,4 +1,0 @@
-INSERT INTO pedidos.producto(id, nombre, description_corta, description_larga, precio_unitario)
- VALUES('89efb206-2aa6-4e21-8a23-5765e3de1f31', "Jamón y morrones", "Pizza de jamón y morrones", "Mozzarella, jamón, morrones y aceitunas verdes", 550.00);
-INSERT INTO pedidos.producto(id, nombre, description_corta, description_larga, precio_unitario)
- VALUES('e29ebd0c-39d2-4054-b0f4-ed2d0ea089a1', "Palmitos", "Pizza de Palmitos", "Mozzarella, palmitos y aceitunas verdes", 600.00);
