+++ src/test/integrationTest/integratioUtilities/BaseTestCase.java
@@ -1,4 +1,7 @@
 package integrationTest.integratioUtilities;
+/**
+ *
+ */
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
